shiftef reslvrs code to index.js
